requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
black>=24.0.0
pyinstaller>=6.0.0